feat(app,server): Add generic HTTP methods and enhance customer update endpoint
- Add generic HTTP methods (get, post, put, delete) to ApiService for flexible API calls
- Enhance updateLPGCustomer endpoint to support both camelCase and snake_case field names
- Add support for additional customer fields: alternatePhone, businessName, gstNumber, creditLimit, preferredCylinderCapacity
- Implement automatic timestamp updates on customer modifications
- Fix type casting in business insights revenue calculation for proper double conversion
- Improve API consistency by accepting multiple naming conventions for request payloads

+++ app/lib/screens/insights/business_insights_screen.dart
@@ -332,7 +332,7 @@ class _BusinessInsightsScreenState extends State<BusinessInsightsScreen> {
                 children: recentTrends.map((day) {
                   final revenue = (day['revenue'] as num).toDouble();
                   final date = DateTime.parse(day['date']);
-                  final height = maxRevenue > 0 ? (revenue / maxRevenue * 150) : 0;
+                  final height = maxRevenue > 0 ? (revenue / maxRevenue * 150).toDouble() : 0.0;
 
                   return Expanded(
                     child: Padding(
